script to create directory using -p
-p: This option is used with the mkdir command to create parent directories if they don't already exist. It ensures tha intermediate directories are created along with the target directory.

For example, when we execute mkdir -p welcome/to/school, it will create the directory welcome/, and within it, it will create the directory to/, and within to/, it will create the directory school/.

If we did not use the -p option and tried to create welcome/to/school directly without the parent directories existing, it would result in an error. The -p option saves us from having to manually create each parent directory separately and handles the creation of the entire directory structure in one go.