script to switch user to betty
Explanation:

su: The su command stands for "substitute user" and is used to switch to another user account. -: The hyphen (-) option passed to su ensures that the new user's environment is loaded as if they had logged in directly.
betty: This is the username of the user we want to switch to. By running this script, it will prompt for the password of the "betty" user and then switch the current user to "betty". The script uses exactly 8 characters for the command (su - betty) as required, plus an additional character for the new line at the end.

Make sure to save the script with a .sh extension (e.g., switch_user.sh), and then you can run it using bash switch_user.sh.